[Refactor]#2 Result클래스를 domain.model 패키지로 이동
