+++ app/src/main/java/com/gallery/kakaogallery/data/repository/ImageRepository.kt
@@ -1,8 +1,8 @@
 package com.gallery.kakaogallery.data.repository
 
 import android.util.Log
-import com.gallery.kakaogallery.Result
-import com.gallery.kakaogallery.ResultError
+import com.gallery.kakaogallery.domain.model.Result
+import com.gallery.kakaogallery.domain.model.ResultError
 import com.gallery.kakaogallery.SearchConstant
 import com.gallery.kakaogallery.data.SaveImageStorage
 import com.gallery.kakaogallery.data.remote.request.ImageSearchRequest
